WIP: First try at porting color opacity widget
This is needed to see the populations of values in cubes, and modify the
color opacity map for volume rendering. It is not yet hooked up, but it
has picked apart some of the dependencies. Rescaling is delicate, and the
VTK widget needs to expose the active volume, active opacity, annd active
color maps so that they can be modified. Signals/slots, other things too.